more auto mode work and refactor
- add alt and speed to gps struct (more work for me later yayyyy!!11!)
- DIDNT fix cmake just made it prettier :))
- change pid lib to use doubles so it's more suitable for gps things
- fix some weird include error
- more progress on auto mode
- the new file (flight.c) is to keep from repeating myself
- it has the code to check flight envelope and run basic PIDs that are used in multiple modes
- add template to make a holding pattern mode
- shoot I need to add throttle control
- create more for me to do
- restructure of normal and tune modes
- and other misc code cleanup!

+++ src/lib/pid.c
@@ -47,18 +47,18 @@ void pid_init(PIDController *pid) {
 
 }
 
-void pid_update(PIDController *pid, float setpoint, float measurement) {
+void pid_update(PIDController *pid, double setpoint, double measurement) {
 
 	/*
 	* Error signal
 	*/
-    float error = setpoint - measurement;
+    double error = setpoint - measurement;
 
 
 	/*
 	* Proportional
 	*/
-    float proportional = pid->Kp * error;
+    double proportional = pid->Kp * error;
 
 
 	/*
